refactor: 에러 핸들러 파일 병합 `/common/exception/*` -> `/global/exception`   - CustomException 위치 변경   - GlobalExceptionHandler 통합
