--- backend/src/main/java/com/tamnara/backend/common/exception/GlobalExceptionHandler.java
@@ -1,29 +1,0 @@
-package com.tamnara.backend.common.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.util.Map;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(CustomException.class)
-    public ResponseEntity<?> handleCustomException(CustomException e) {
-        return ResponseEntity.status(e.getStatus()).body(Map.of(
-                "success", false,
-                "message", e.getMessage()
-        ));
-    }
-
-    // Optional: 일반 예외 처리
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<?> handleGenericException(Exception e) {
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(Map.of(
-                "success", false,
-                "message", "서버에 문제가 발생했습니다."
-        ));
-    }
-}
